feat: 리뷰 삭제 서비스단 메서드 구현 ( #9 )

+++ re-place/src/main/java/com/replace/re/place/category/service/CategoryService.java
@@ -4,6 +4,8 @@ import com.replace.re.place.category.dao.CategoryDao;
 import com.replace.re.place.category.dto.CategoryDto;
 import com.replace.re.place.global.ErrorCode;
 import com.replace.re.place.global.exception.category.CategoryNotCreatedException;
+import com.replace.re.place.global.exception.category.CategoryNotDeletedException;
+import com.replace.re.place.global.exception.category.CategoryNotFoundException;
 import com.replace.re.place.global.exception.category.ReviewCategoryNotCreatedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -74,18 +76,34 @@ public class CategoryService {
     // reviewId를 가지는 review의 카테고리 전부 삭제.
     public Boolean deleteReviewCategory(Long reviewId){
 
-        List<CategoryDto> categoryDtos = categoryDao.getCategoriesByReviewId(reviewId);
 
-        Boolean isDeleted = categoryDao.deleteReviewCategoryByReviewId(reviewId);
-        if(isDeleted){
+        Boolean isReviewCategoryExist = categoryDao.checkReviewCategoryExistByReviewId(reviewId);
+
+        //리뷰에 카테고리가 한 개도 없을 수 있기 때문에, 리뷰 카테고리가 없더라도 예외발생 X.
+        if(isReviewCategoryExist) {
+            List<CategoryDto> categoryDtos = categoryDao.getCategoriesByReviewId(reviewId);
+
+            categoryDao.deleteReviewCategoryByReviewId(reviewId);
+
+
             Iterator<CategoryDto> it = categoryDtos.iterator();
-            while(it.hasNext()){
+            while (it.hasNext()) {
                 CategoryDto categoryDto = it.next();
                 Long categoryId = categoryDto.getCategoryId();
 
-                if(categoryDao.checkReviewCategoryExistByCategoryId(categoryId)){    // 삭제한 리뷰 내에서만 쓰이는 카테고리인 경우
-                    categoryDao.deleteCategory(categoryId);
+                Boolean isCategoryExist = categoryDao.checkCategoryExist(categoryId);
+                if(isCategoryExist) {
+                    if (categoryDao.checkReviewCategoryExistByCategoryId(categoryId)) {    // 삭제한 리뷰 내에서만 쓰이는 카테고리인 경우
+
+                        categoryDao.deleteCategory(categoryId);
+                        Boolean isCategoryDeleted = !categoryDao.checkCategoryExist(categoryId);
+                        if(!isCategoryDeleted){
+                            throw new CategoryNotDeletedException(CATEGORY_NOT_DELETED);
+                        }
+                    }
                 }
+
+                throw new CategoryNotFoundException(CATEGORY_NOT_FOUND);
             }
         }
